TTD add use top most location regression test
