+++ modules/ttdBidAdapter.js
@@ -133,7 +133,7 @@ function getUser(bidderRequest) {
 
 function getSite(bidderRequest, firstPartyData) {
   var site = utils.mergeDeep({
-    page: utils.deepAccess(bidderRequest, 'refererInfo.topmostLocation') || utils.deepAccess(bidderRequest, 'refererInfo.page'),
+    page: utils.deepAccess(bidderRequest, 'refererInfo.page'),
     ref: utils.deepAccess(bidderRequest, 'refererInfo.ref'),
     publisher: {
       id: utils.deepAccess(bidderRequest, 'bids.0.params.publisherId'),
